requirements.txt
ipykernel
Flask
cdlib
leidenalg
networkx
matplotlib
pandas
scikit-learn
cdlib
community
numpy
bayanpy
pyclustering
ASLPAw
wurlitzer
waitress
requests
python-dotenv
Flask-Cors
dill